shore: dto and entity refactoring
