+++ src/core/Product/domain/product-types.ts
@@ -1,0 +1,10 @@
+export type ProductDto = {
+	ean: string
+	name: string
+	description?: string
+	cost: number
+}
+
+export type CreateProductDto = ProductDto
+
+export type UpdateProductDto = Partial<CreateProductDto>
