--- src/core/Product/dto/index.ts
@@ -1,3 +1,0 @@
-export * from './create-product.dto'
-export * from './update-product.dto'
-export * from './product.dto'
