+++ src/core/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 import { UserService } from './user.service'
 import { UserGQLResolver } from './gql/user-gql-resolver'
-import { PrismaModule } from 'src/infrastructure/prisma/prisma.module'
-import { PrismaService } from 'src/infrastructure/prisma/prisma.service'
+import { PrismaModule } from '../../infrastructure/prisma/prisma.module'
+import { PrismaService } from '../../infrastructure/prisma/prisma.service'
 import { AbstractUserRepository } from './repository/abstract.user.repository'
 import { UserRepository } from './repository/user.repository'
